refactored the code based on new insights
