+++ src/helper.js
@@ -1,0 +1,4 @@
+const Employee = require('../lib/Employee.js')
+const Engineer = require('../lib/Engineer.js')
+const Intern = require('../lib/Intern.js')
+const Manager = require('../lib/Manager.js')
